requirements.txt
openai
python-dotenv

langchain
langchain-openai
langgraph
langchain-community
faiss-cpu

fastapi
uvicorn[standard]
pydantic

# Optional but nice to have for local testing / audio utils
pydub

python-multipart
